+++ src/main/java/com/sprint/mission/discodeit/service/AuthService.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.auth.service;
+package com.sprint.mission.discodeit.service;
 
 import com.sprint.mission.discodeit.dto.data.UserDto;
 import com.sprint.mission.discodeit.dto.request.UserRoleUpdateRequest;
